+++ eframe-core/src/main/java/net/linlan/frame/comm/security/UserDetailsServiceImpl.java
@@ -75,6 +75,6 @@ public class UserDetailsServiceImpl implements UserDetailsService {
 
     public UserDetails createLoginUser(AdminUserDto user) {
         return new FrameAdminUser(user.getId(), user.getDeptId(), user.getUsername(),
-            user.getPassword(), user.getLastAppId(), sysPermissionService.getMenuPermission(user));
+            user.getPassword(), user.getLastAppId(), user.getAdminType(), sysPermissionService.getMenuPermission(user));
     }
 }